Automate reference point selection
Add code to update the reference point by randomly sampling points around the cursor when zooming. Since users are likely to zoom in towards areas of interest, and such areas are often near the border of the set, it is assumed that there will be points in the set around the cursor.

This method can fail to find suitable reference points when zooming towards sparse areas of the set, since its ability to find suitable points is proportional to the amount of space around the cursor which is in the set.

In future, better methods of finding reference points and ways of compensating for poor reference points will be needed.